🐛 ShareButton Atom IconButton으로 변경
공유, 삭제, 추가 등 이미지 버튼이 여러 개 쓰일 것 같다고 예상되어 img src와 click handler를 받는 버튼으로 수정. img와 button으로 atom 분리 후 단계 업그레이드 고려 중
