+++ client/src/components/atoms/index.tsx
@@ -1,5 +1,5 @@
 export { default as BoxButton } from 'components/atoms/BoxButton';
 export { default as Input } from 'components/atoms/Input';
-export { default as ShareButton } from 'components/atoms/ShareButton';
+export { default as IconButton } from 'components/atoms/IconButton';
 export { default as Toast } from 'components/atoms/Toast';
 export { default as Title } from 'components/atoms/Title';
